Fixes some warnings uncovered by -Wextra

+++ src/idmef-tree-wrap.c
@@ -265,7 +265,7 @@ static void list_insert(prelude_list_t *head, prelude_list_t *item, int pos)
  */
 idmef_additional_data_type_t idmef_additional_data_type_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_additional_data_type_t val;
               const char *name;
@@ -320,7 +320,7 @@ const char *idmef_additional_data_type_to_string(idmef_additional_data_type_t va
                 { IDMEF_ADDITIONAL_DATA_TYPE_XML, "xml" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -345,7 +345,7 @@ struct idmef_additional_data {
  */
 idmef_reference_origin_t idmef_reference_origin_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_reference_origin_t val;
               const char *name;
@@ -390,7 +390,7 @@ const char *idmef_reference_origin_to_string(idmef_reference_origin_t val)
                 { IDMEF_REFERENCE_ORIGIN_OSVDB, "osvdb" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -428,7 +428,7 @@ struct idmef_classification {
  */
 idmef_user_id_type_t idmef_user_id_type_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_user_id_type_t val;
               const char *name;
@@ -475,7 +475,7 @@ const char *idmef_user_id_type_to_string(idmef_user_id_type_t val)
                 { IDMEF_USER_ID_TYPE_OTHER_PRIVS, "other-privs" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -502,7 +502,7 @@ struct idmef_user_id {
  */
 idmef_user_category_t idmef_user_category_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_user_category_t val;
               const char *name;
@@ -541,7 +541,7 @@ const char *idmef_user_category_to_string(idmef_user_category_t val)
                 { IDMEF_USER_CATEGORY_OS_DEVICE, "os-device" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -565,7 +565,7 @@ struct idmef_user {
  */
 idmef_address_category_t idmef_address_category_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_address_category_t val;
               const char *name;
@@ -628,7 +628,7 @@ const char *idmef_address_category_to_string(idmef_address_category_t val)
                 { IDMEF_ADDRESS_CATEGORY_IPV6_NET_MASK, "ipv6-net-mask" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -701,7 +701,7 @@ struct idmef_snmp_service {
  */
 idmef_service_type_t idmef_service_type_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_service_type_t val;
               const char *name;
@@ -740,7 +740,7 @@ const char *idmef_service_type_to_string(idmef_service_type_t val)
                 { IDMEF_SERVICE_TYPE_SNMP, "snmp" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -777,7 +777,7 @@ struct idmef_service {
  */
 idmef_node_category_t idmef_node_category_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_node_category_t val;
               const char *name;
@@ -836,7 +836,7 @@ const char *idmef_node_category_to_string(idmef_node_category_t val)
                 { IDMEF_NODE_CATEGORY_WFW, "wfw" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -862,7 +862,7 @@ struct idmef_node {
  */
 idmef_source_spoofed_t idmef_source_spoofed_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_source_spoofed_t val;
               const char *name;
@@ -901,7 +901,7 @@ const char *idmef_source_spoofed_to_string(idmef_source_spoofed_t val)
                 { IDMEF_SOURCE_SPOOFED_NO, "no" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -957,7 +957,7 @@ struct idmef_inode {
  */
 idmef_checksum_algorithm_t idmef_checksum_algorithm_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_checksum_algorithm_t val;
               const char *name;
@@ -1010,7 +1010,7 @@ const char *idmef_checksum_algorithm_to_string(idmef_checksum_algorithm_t val)
                 { IDMEF_CHECKSUM_ALGORITHM_GOST, "Gost" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1035,7 +1035,7 @@ struct idmef_checksum {
  */
 idmef_file_category_t idmef_file_category_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_file_category_t val;
               const char *name;
@@ -1072,7 +1072,7 @@ const char *idmef_file_category_to_string(idmef_file_category_t val)
                 { IDMEF_FILE_CATEGORY_ORIGINAL, "original" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1086,7 +1086,7 @@ const char *idmef_file_category_to_string(idmef_file_category_t val)
  */
 idmef_file_fstype_t idmef_file_fstype_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_file_fstype_t val;
               const char *name;
@@ -1139,7 +1139,7 @@ const char *idmef_file_fstype_to_string(idmef_file_fstype_t val)
                 { IDMEF_FILE_FSTYPE_ISO9660, "iso9660" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1184,7 +1184,7 @@ struct idmef_file {
  */
 idmef_linkage_category_t idmef_linkage_category_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_linkage_category_t val;
               const char *name;
@@ -1229,7 +1229,7 @@ const char *idmef_linkage_category_to_string(idmef_linkage_category_t val)
                 { IDMEF_LINKAGE_CATEGORY_SYMBOLIC_LINK, "symbolic-link" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1256,7 +1256,7 @@ struct idmef_linkage {
  */
 idmef_target_decoy_t idmef_target_decoy_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_target_decoy_t val;
               const char *name;
@@ -1295,7 +1295,7 @@ const char *idmef_target_decoy_to_string(idmef_target_decoy_t val)
                 { IDMEF_TARGET_DECOY_NO, "no" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1360,7 +1360,7 @@ struct idmef_alertident {
  */
 idmef_impact_severity_t idmef_impact_severity_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_impact_severity_t val;
               const char *name;
@@ -1401,7 +1401,7 @@ const char *idmef_impact_severity_to_string(idmef_impact_severity_t val)
                 { IDMEF_IMPACT_SEVERITY_HIGH, "high" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1415,7 +1415,7 @@ const char *idmef_impact_severity_to_string(idmef_impact_severity_t val)
  */
 idmef_impact_completion_t idmef_impact_completion_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_impact_completion_t val;
               const char *name;
@@ -1452,7 +1452,7 @@ const char *idmef_impact_completion_to_string(idmef_impact_completion_t val)
                 { IDMEF_IMPACT_COMPLETION_SUCCEEDED, "succeeded" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1466,7 +1466,7 @@ const char *idmef_impact_completion_to_string(idmef_impact_completion_t val)
  */
 idmef_impact_type_t idmef_impact_type_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_impact_type_t val;
               const char *name;
@@ -1511,7 +1511,7 @@ const char *idmef_impact_type_to_string(idmef_impact_type_t val)
                 { IDMEF_IMPACT_TYPE_USER, "user" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1537,7 +1537,7 @@ struct idmef_impact {
  */
 idmef_action_category_t idmef_action_category_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_action_category_t val;
               const char *name;
@@ -1578,7 +1578,7 @@ const char *idmef_action_category_to_string(idmef_action_category_t val)
                 { IDMEF_ACTION_CATEGORY_TAKEN_OFFLINE, "taken-offline" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1603,7 +1603,7 @@ struct idmef_action {
  */
 idmef_confidence_rating_t idmef_confidence_rating_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_confidence_rating_t val;
               const char *name;
@@ -1644,7 +1644,7 @@ const char *idmef_confidence_rating_to_string(idmef_confidence_rating_t val)
                 { IDMEF_CONFIDENCE_RATING_HIGH, "high" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1711,7 +1711,7 @@ struct idmef_overflow_alert {
  */
 idmef_alert_type_t idmef_alert_type_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_alert_type_t val;
               const char *name;
@@ -1752,7 +1752,7 @@ const char *idmef_alert_type_to_string(idmef_alert_type_t val)
                 { IDMEF_ALERT_TYPE_OVERFLOW, "overflow" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
@@ -1811,7 +1811,7 @@ struct idmef_heartbeat {
  */
 idmef_message_type_t idmef_message_type_to_numeric(const char *name)
 {
-        int i;
+        size_t i;
         const struct {
               idmef_message_type_t val;
               const char *name;
@@ -1848,7 +1848,7 @@ const char *idmef_message_type_to_string(idmef_message_type_t val)
                 { IDMEF_MESSAGE_TYPE_HEARTBEAT, "heartbeat" },
         };
 
-        if ( val < 0 || val >= (sizeof(tbl) / sizeof(*tbl)) )
+        if ( val < 0 || (size_t) val >= (sizeof(tbl) / sizeof(*tbl)) )
                 return NULL;
 
         return tbl[val].name;
